requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
boto3==1.26.141
botocore==1.29.141
CacheControl==0.13.1
cachetools==5.3.1
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
Django==4.2.1
django-cors-headers==4.1.0
django-jazzmin==2.6.0
django-storages==1.13.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
Faker==18.10.0
fcm-django==2.0.0
firebase-admin==6.1.0
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kombu==5.2.4
msgpack==1.0.5
multidict==6.0.4
Pillow==9.5.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
rsa==4.9
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
twilio==8.8.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
zope.interface==6.0
